// vue3/src/openapi/models/RecipeRequest.ts
export interface RecipeRequest {
    /**
     * 
     * @type {string}
     * @memberof RecipeRequest
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof RecipeRequest
     */
    description?: string | null;
    /**
     * 
     * @type {Array<KeywordRequest>}
     * @memberof RecipeRequest
     */
    keywords?: Array<KeywordRequest>;
    /**
     * 
     * @type {Array<StepRequest>}
     * @memberof RecipeRequest
     */
    steps: Array<StepRequest>;
    /**
     * 
     * @type {number}
     * @memberof RecipeRequest
     */
    workingTime?: number;
    /**
     * 
     * @type {number}
     * @memberof RecipeRequest
     */
    waitingTime?: number;
    /**
     * 
     * @type {string}
     * @memberof RecipeRequest
     */
    sourceUrl?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof RecipeRequest
     */
    internal?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof RecipeRequest
     */
    showIngredientOverview?: boolean;
    /**
     * 
     * @type {NutritionInformationRequest}
     * @memberof RecipeRequest
     */
    nutrition?: NutritionInformationRequest | null;
    /**
     * 
     * @type {Array<PropertyRequest>}
     * @memberof RecipeRequest
     */
    properties?: Array<PropertyRequest>;
    /**
     * 
     * @type {number}
     * @memberof RecipeRequest
     */
    servings?: number;
    /**
     * 
     * @type {string}
     * @memberof RecipeRequest
     */
    servingsText?: string;
    /**
     * 
     * @type {number}
     * @memberof RecipeRequest
     */
    diameter?: number;
    /**
     * 
     * @type {string}
     * @memberof RecipeRequest
     */
    diameterText?: string;
    /**
     * 
     * @type {boolean}
     * @memberof RecipeRequest
     */
    _private?: boolean;
    /**
     * 
     * @type {Array<UserRequest>}
     * @memberof RecipeRequest
     */
    shared?: Array<UserRequest>;
    /**
     * 
     * @type {number}
     * @memberof RecipeRequest
     */
    id?: number;
}