// vue3/src/openapi/models/SourceImportRecipe.ts
export interface SourceImportRecipe {
    /**
     * 
     * @type {Array<SourceImportStep>}
     * @memberof SourceImportRecipe
     */
    steps: Array<SourceImportStep>;
    /**
     * 
     * @type {boolean}
     * @memberof SourceImportRecipe
     */
    internal?: boolean;
    /**
     * 
     * @type {string}
     * @memberof SourceImportRecipe
     */
    sourceUrl: string;
    /**
     * 
     * @type {string}
     * @memberof SourceImportRecipe
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof SourceImportRecipe
     */
    description?: string;
    /**
     * 
     * @type {number}
     * @memberof SourceImportRecipe
     */
    servings?: number;
    /**
     * 
     * @type {string}
     * @memberof SourceImportRecipe
     */
    servingsText?: string;
    /**
     * 
     * @type {number}
     * @memberof SourceImportRecipe
     */
    workingTime?: number;
    /**
     * 
     * @type {number}
     * @memberof SourceImportRecipe
     */
    waitingTime?: number;
    /**
     * 
     * @type {string}
     * @memberof SourceImportRecipe
     */
    imageUrl?: string;
    /**
     * 
     * @type {Array<SourceImportKeyword>}
     * @memberof SourceImportRecipe
     */
    keywords?: Array<SourceImportKeyword>;
    /**
     * 
     * @type {Array<SourceImportProperty>}
     * @memberof SourceImportRecipe
     */
    properties?: Array<SourceImportProperty>;
}